Put in place the basics of an API between the flask app and the maths app.
For this, a relationship module has been introduced to document the heirarchy of questions.

For example, we have question modules structured like so now:

@relationships.root
class RootPart(relationships.QuestionPart):
    #
    #
    #

@relationships.is_child_of(RootPart)
class SomeChild(relationships.QuestionPart):
    #
    #
    #

Which allows us to programmatically build the question rather than add parts manually.

We also have the new api package which will serve as the go-between. It's all basic for now, but it's something.