Introduce a new class method for our questions to store all sets of valid question parameters.
The method "store_question" uses a generator that is implemented
in the QuestionPart ("enumerate_questions").

The storage is optimised for fast *access*. Writing all the questions
is quite slow, since we have to iterate over many possibilities. However,
once all questions are written, we know the start and end bytes of each
question, so we merely need to choose a random question and then
we can read it straight away.